% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toplevel.R
\name{infer.rmm}
\alias{infer.rmm}
\title{Inference calculations for RMMs}
\usage{
infer.rmm(r)
}
\arguments{
\item{model}{A fitted recurrent multinomial model of class 'rmm'}
}
\value{
A fitted recurrent multinomial model of class 'rmm' with standard
errors eta.se, beta.se, zeta.se for eta, beta, and zeta respectively
}
\description{
Calculate the standard errors of coefficients from the
Hessian matrix if not done automatically in \code{rmm}, i.e. when
\code{hessian = FALSE}
}
