% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toplevel.R
\name{rmm}
\alias{rmm}
\alias{predict.rmm}
\title{Estimate a Recurrent Multinomial Model}
\usage{
rmm(
  sequences,
  D,
  prob.terms = c(),
  recur.terms = c(),
  recur.seq.covar = list(),
  hessian = TRUE,
  ...
)

\method{predict}{rmm}(
  model,
  newdata,
  prob.terms = c(),
  recur.terms = c(),
  recur.seq.covar = list(),
  lambda = NA,
  ...
)
}
\arguments{
\item{sequences}{A list of vectors with sequence data}

\item{prob.terms}{A list of predefined covariate terms to include in the X
matrix predicting baseline element-specific probability. See details.}

\item{recur.terms}{A list of predefined covariate terms to include in the Z
matrix predicting recurrence coefficients. See details.}

\item{hessian}{Logical value indicating whether the hessian should be
computed}

\item{lambda, lambda2}{Regularization parameter for recurrence and baseline
probability coefficients respectively}

\item{recurr.seq.covar}{A list of functions for defining custom recurrence
covariates}

\item{etastart, betastart, zetastart}{Starting values for optimization.
Defaults to 0.}

\item{parallel}{Logical values indicating whether to use parallel computation}

\item{cores}{Number of cores to use in parallel computation}

\item{num.grad}{Logical value indicating whether to use a numerical gradient}
}
\value{
A fitted recurrent multinomial model of class 'rmm'
}
\description{
\code{RMM} estimates a recurrent mutlinomial model of specified length with
covariate terms.
}
\details{
This is the primary estimation function for fitting RMMs
}
\section{Methods (by generic)}{
\itemize{
\item \code{predict}: Evaluate the probability of observed sequences on a prefit
rmm model
}}

\section{Baseline Covariates}{

These covariates provide an element specific estimate for the probability
of an element appearing, as with a normal multinomial model coefficient.
Prespecified baseline covariates include:
\itemize{
\item \code{'constant'} - Intercept
\item \code{'sequence'} - Sequence specific probability
\item \code{'sequence fraction'} - Fraction of sequence complete
\item \code{'presence'} - Whether the element occurs at all in the observed
sequence
}
}

\section{Recurrence Covariates}{

These covariates provide an lag specific estimate for recurrence of a
previously observed element reappearing.
Prespecified recurrence covariates include:
\itemize{
\item \code{'constant'} - Intercept
\item \code{'sequence'} - Sequence specific probability
\item \code{'sequence fraction'} - Fraction of sequence complete
\item \code{'sequence fraction squared'} - (Fraction of sequence complete)^2
}
}

\section{Custom Covariates}{

Custom covariates can be defined by adding functions to the
\code{recur.seq.covar} list. These functions should take four parameters:
\code{s0}, \code{i}, \code{l}, and \code{a} where \code{s0} is a observed
element in at position \code{l} in the \code{i}th sequence and \code{a} is a
possible element.

Possible uses for these functions include controlling for elements
belonging to a certain category, sequences of a certain type, timing
specific recurrences, or combinations of those effects.
}

